SQL Queries Join, Navigation Buttons (Backlog), Page Help
Used Joins in SQL queries to reduce code and database usage.
The buttons allow users to get around the site without going through the
navigation in the banner.
Help has been added in a collapse at the top of each page with
information on how to use specific aspects of functionality.